Split all clean function into `find_*` vs `rm_*`

+++ conda/cli/main_clean.py
@@ -3,13 +3,13 @@
 # SPDX-License-Identifier: BSD-3-Clause
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import fnmatch
 from logging import getLogger
-from os import listdir, lstat, unlink, walk
-from os.path import exists, getsize, isdir, join
+from os import listdir, lstat, walk
+from os.path import getsize, isdir, join
+from typing import Iterable, Tuple
 import sys
 
-from ..base.constants import CONDA_PACKAGE_EXTENSIONS, CONDA_TEMP_EXTENSION
+from ..base.constants import CONDA_PACKAGE_EXTENSIONS, CONDA_TEMP_EXTENSIONS
 from ..base.context import context
 
 log = getLogger(__name__)
@@ -17,14 +17,9 @@ _EXTS = (*CONDA_PACKAGE_EXTENSIONS, *(f"{e}.part" for e in CONDA_PACKAGE_EXTENSI
 
 
 def find_tarballs():
-    from ..core.package_cache_data import PackageCacheData
-
     pkgs_dirs = {}
     total_size = 0
-    for package_cache in PackageCacheData.writable_caches(context.pkgs_dirs):
-        pkgs_dir = package_cache.pkgs_dir
-        if not isdir(pkgs_dir):
-            continue
+    for pkgs_dir in find_pkgs_dirs():
         root, _, filenames = next(walk(pkgs_dir))
         for fn in filenames:
             if fn.endswith(_EXTS):
@@ -32,7 +27,6 @@ def find_tarballs():
                 total_size += getsize(join(root, fn))
 
     return {"pkgs_dirs": pkgs_dirs, "total_size": total_size}
-
 
 def rm_tarballs(args, pkgs_dirs, total_size, verbose=True):
     from .common import confirm_yn
@@ -84,17 +78,12 @@ def rm_tarballs(args, pkgs_dirs, total_size, verbose=True):
                 else:
                     log.info("%r", e)
 
-
 def find_pkgs():
     # TODO: This doesn't handle packages that have hard links to files within
     # themselves, like bin/python3.3 and bin/python3.3m in the Python package
     warnings = []
     pkgs_dirs = {}
-    for pkgs_dir in context.pkgs_dirs:
-        if not exists(pkgs_dir):
-            if not context.json:
-                print("WARNING: {0} does not exist".format(pkgs_dir))
-            continue
+    for pkgs_dir in find_pkgs_dirs():
         pkgs = [i for i in listdir(pkgs_dir) if isdir(join(pkgs_dir, i, 'info'))]
         for pkg in pkgs:
             breakit = False
@@ -136,7 +125,6 @@ def find_pkgs():
         "pkg_sizes": pkg_sizes,
     }
 
-
 def rm_pkgs(args, pkgs_dirs, warnings, total_size, pkg_sizes, verbose=True):
     from .common import confirm_yn
     from ..gateways.disk.delete import rm_rf
@@ -177,59 +165,78 @@ def rm_pkgs(args, pkgs_dirs, warnings, total_size, pkg_sizes, verbose=True):
                 print("removing %s" % pkg)
             rm_rf(join(pkgs_dir, pkg))
 
+def find_index_cache() -> Tuple[str]:
+    files = []
+    for pkgs_dir in find_pkgs_dirs():
+        # caches are directories in pkgs_dir
+        path = join(pkgs_dir, "cache")
+        if isdir(path):
+            files.append(path)
+    return files
 
-def rm_index_cache():
-    from ..gateways.disk.delete import rm_rf
+def find_pkgs_dirs() -> Tuple[str]:
     from ..core.package_cache_data import PackageCacheData
-    for package_cache in PackageCacheData.writable_caches():
-        rm_rf(join(package_cache.pkgs_dir, 'cache'))
+
+    return tuple(pc.pkgs_dir for pc in PackageCacheData.writable_caches() if isdir(pc.pkgs_dir))
 
 
-def rm_rf_pkgs_dirs():
+def find_tempfiles(paths: Iterable[str]) -> Tuple[str]:
+    tempfiles = []
+    for path in sorted(set(paths or [sys.prefix])):
+        # tempfiles are files in path
+        for (
+            root,
+            _,
+            files,
+        ) in walk(path):
+            for file in files:
+                # tempfiles also end in .c~ or .trash
+                if not file.endswith(CONDA_TEMP_EXTENSIONS):
+                    continue
+
+                tempfiles.append(join(root, file))
+
+    return tempfiles
+
+
+def rm_items(args, items: Tuple[str], verbose: bool, name: str) -> None:
     from .common import confirm_yn
-    from ..common.io import dashlist
     from ..gateways.disk.delete import rm_rf
-    from ..core.package_cache_data import PackageCacheData
 
-    writable_pkgs_dirs = tuple(
-        pc.pkgs_dir for pc in PackageCacheData.writable_caches() if isdir(pc.pkgs_dir)
-    )
+    if not items:
+        if verbose:
+            print(f"There are no {name} to remove.")
+        return
+
+    if verbose:
+        if args.verbosity:
+            print(f"Will remove the following {name}:")
+            for item in items:
+                print(f"  - {item}")
+            print()
+        else:
+            print(f"Will remove {len(items)} {name}.")
+
+    if args.dry_run:
+        return
     if not context.json or not context.always_yes:
-        print("Remove all contents from the following package caches?%s"
-              % dashlist(writable_pkgs_dirs))
         confirm_yn()
 
-    for pkgs_dir in writable_pkgs_dirs:
-        rm_rf(pkgs_dir)
-
-    return writable_pkgs_dirs
-
-
-def clean_tmp_files(path=None):
-    if not path:
-        path = sys.prefix
-    for root, dirs, fns in walk(path):
-        for fn in fns:
-            if (fnmatch.fnmatch(fn, "*.trash") or
-                    fnmatch.fnmatch(fn, "*" + CONDA_TEMP_EXTENSION)):
-                file_path = join(root, fn)
-                try:
-                    unlink(file_path)
-                except EnvironmentError:
-                    log.warn("File at {} could not be cleaned up.  "
-                             "It's probably still in-use.".format(file_path))
+    for item in items:
+        rm_rf(item)
 
 def _execute(args, parser):
     json_result = {"success": True}
+    verbose = not (context.json or context.quiet)
 
-    if args.source_cache:
+    if verbose and args.source_cache:
         print("WARNING: 'conda clean --source-cache' is deprecated.\n"
               "    Use 'conda build purge-all' to remove source cache files.",
               file=sys.stderr)
 
     if args.force_pkgs_dirs:
-        writable_pkgs_dirs = rm_rf_pkgs_dirs()
-        json_result['pkgs_dirs'] = writable_pkgs_dirs
+        json_result["pkgs_dirs"] = pkgs_dirs = find_pkgs_dirs()
+        rm_items(args, pkgs_dirs, verbose=verbose, name="package cache(s)")
 
         # we return here because all other clean operations target individual parts of
         # package caches
@@ -242,26 +249,22 @@ def _execute(args, parser):
 
     if args.tarballs or args.all:
         json_result["tarballs"] = tars = find_tarballs()
-        rm_tarballs(args, **tars, verbose=not (context.json or context.quiet))
+        rm_tarballs(args, **tars, verbose=verbose)
 
     if args.index_cache or args.all:
-        json_result['index_cache'] = {
-            'files': [join(context.pkgs_dirs[0], 'cache')]
-        }
-        rm_index_cache()
+        cache = find_index_cache()
+        json_result["index_cache"] = {"files": cache}
+        rm_items(args, cache, verbose=verbose, name="index cache(s)")
 
     if args.packages or args.all:
         json_result["packages"] = pkgs = find_pkgs()
-        rm_pkgs(args, **pkgs, verbose=not (context.json or context.quiet))
+        rm_pkgs(args, **pkgs, verbose=verbose)
 
-    if args.all:
-        clean_tmp_files(sys.prefix)
-    elif args.tempfiles:
-        for path in args.tempfiles:
-            clean_tmp_files(path)
+    if args.tempfiles or args.all:
+        tmps = find_tempfiles(args.tempfiles)
+        rm_items(args, tmps, verbose=verbose, name="tempfile(s)")
 
     return json_result
-
 
 def execute(args, parser):
     from .common import stdout_json
